detection: better propagation of auto estimate as midi events
- set current estimate to zero if silent is detected or if the
  pitch estimation fails
- rename current to current_note (clearer)
- estimate velocity based on the energy

+++ frontend/rust/src/api/event.rs
@@ -3,7 +3,7 @@ pub use backend::event::{MidiEvent, Status};
 #[flutter_rust_bridge::frb(mirror(MidiEvent))]
 pub struct _MidiEvent {
     pub status: Status,
-    pub velocity: u32,
+    pub velocity: u8,
     pub note: u8,
     pub note_name: String,
     pub raw: Vec<u8>,
